# openapi/sagewai.yaml
openapi: 3.1.0
info:
  title: Sagewai API
  version: 0.1.0
  description: REST API for Sagewai agent infrastructure platform.
  contact:
    email: hello@sagewai.ai
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8100
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key or JWT token
  schemas:
...
    AgentDetail:
      type: object
      properties:
        name: { type: string }
        capabilities: { type: array, items: { type: string } }
        model: { type: string }
        system_prompt: { type: string }
        max_iterations: { type: integer }
        tools: { type: array, items: { type: string } }
        total_runs: { type: integer }
        strategy: { type: string }
        tags: { type: array, items: { type: string } }
    RunSummary:
      type: object
      properties:
        run_id: { type: string }
        agent_name: { type: string }
        status: { type: string }
        input_preview: { type: string }
        output_preview: { type: string }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        total_tokens: { type: integer }
        run_type: { type: string }
    RunDetail:
      type: object
      properties:
        run_id: { type: string }
        agent_name: { type: string }
        status: { type: string }
        input_text: { type: string }
        output_text: { type: string }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        total_tokens: { type: integer }
        tool_calls: { type: array, items: { type: object } }
        steps: { type: array, items: { type: object } }
    PaginatedResponse:
      type: object
      properties:
        items: { type: array, items: { type: object } }
        next_cursor: { type: string }
        has_more: { type: boolean }
...
paths:
...
  /admin/agents/{agent_name}:
    get:
      operationId: getAgent
      tags: [Admin]
      summary: Agent detail
      parameters:
        - name: agent_name
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Full agent detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentDetail' }
  /admin/agents/{agent_name}/config:
    patch:
      operationId: updateAgentConfig
      tags: [Admin]
      summary: Update agent configuration
      parameters:
        - name: agent_name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model: { type: string }
                system_prompt: { type: string }
                temperature: { type: number }
                max_tokens: { type: integer }
                tags: { type: array, items: { type: string } }
      responses:
        '200':
          description: Updated config
  /admin/runs:
    get:
      operationId: listRuns
      tags: [Admin]
      summary: List runs
      parameters:
        - name: agent_name
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Paginated run list
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PaginatedResponse' }
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/RunSummary' } }
  /admin/runs/{run_id}:
    get:
      operationId: getRun
      tags: [Admin]
      summary: Run detail
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Full run detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunDetail' }
  /admin/runs/{run_id}/pause:
    post:
      operationId: pauseRun
      tags: [Admin]
      summary: Pause a running agent
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Run paused
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id: { type: string }
                  action: { type: string }
                  status: { type: string }
  /admin/runs/{run_id}/resume:
    post:
      operationId: resumeRun
      tags: [Admin]
      summary: Resume a paused agent
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Run resumed
  /admin/runs/{run_id}/cancel:
    post:
      operationId: cancelRun
      tags: [Admin]
      summary: Cancel a running agent
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Run cancelled
  /admin/sessions:
    get:
      operationId: listSessions
      tags: [Admin]
      summary: List active sessions
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Session list
  /admin/sessions/{session_id}:
    get:
      operationId: getSession
      tags: [Admin]
      summary: Session detail
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Session detail

  # ── Health ────────────────────────────────────────────────
...